# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Network Clone",
        "description": "Social Network Clone API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "feed/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get posts.",
                "operationId": "11332c9d31a56a83b8bf5089279d3fb3",
                "responses": {
                    "200": {
                        "description": "Posts",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a post.",
                "operationId": "7cef8f4a5fd845dc1ad5f14d0d6974d2",
                "responses": {
                    "200": {
                        "description": "Post created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/posts/{post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a post.",
                "operationId": "c42a83b04824548a1c0b04db54c395c3",
                "responses": {
                    "200": {
                        "description": "Post",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update a post.",
                "operationId": "b8439704f2430b00e3995bb393e03c69",
                "responses": {
                    "200": {
                        "description": "Post updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post.",
                "operationId": "ecbb243d045a88668c7c39945f2d43b5",
                "responses": {
                    "200": {
                        "description": "Post deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/posts/reposts": {
            "get": {
                "tags": [
                    "Reposts"
                ],
                "summary": "Get user reposts.",
                "operationId": "fe06335a3868c0ec9d3cc81ae4e46516",
                "responses": {
                    "200": {
                        "description": "User reposts",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/posts/{post}/reposts": {
            "get": {
                "tags": [
                    "Reposts"
                ],
                "summary": "Get reposts for a post.",
                "operationId": "455c3374471dc27f2436435ea6e6039f",
                "responses": {
                    "200": {
                        "description": "Post reposts",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reposts"
                ],
                "summary": "Repost a post.",
                "operationId": "1297df9094324a66758d9770bedf6f74",
                "responses": {
                    "200": {
                        "description": "Repost",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/posts/{post}/reposts/{repost}": {
            "delete": {
                "tags": [
                    "Reposts"
                ],
                "summary": "Unrepost a post.",
                "operationId": "2e9f1f15341829fca678a6d461e45f77",
                "responses": {
                    "200": {
                        "description": "Post unreposted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
...
}